requirements.txt
kafka-python==2.0.2
pyspark==3.3.0
pycoingecko==3.1.0
requests==2.31.0 